stock_project/requirements.txt
redis~=4.5.5
kafka~=1.3.5
Django==4.0.2
djangorestframework==3.14.0
flake8==5.0.4
django-redis
flake8==5.0.4
